HCK-12235: move connection logic to separate file
