+++ reverse_engineering/helpers/connections/constants.js
@@ -1,0 +1,29 @@
+const ALREADY_CONNECTED_STATUS = 405502;
+const CANT_REACH_SNOWFLAKE_ERROR_STATUS = 401001;
+const CONNECTION_TIMED_OUT_CODE = 'CONNECTION_TIMED_OUT';
+
+const DEFAULT_CLIENT_APP_ID = 'JavaScript';
+const DEFAULT_CLIENT_APP_VERSION = '1.5.1';
+const DEFAULT_WAREHOUSE = 'COMPUTE_WH';
+const DEFAULT_ROLE = 'PUBLIC';
+const HACKOLADE_APPLICATION = 'Hackolade';
+const CLOUD_PLATFORM_POSTFIXES = ['gcp', 'aws', 'azure'];
+
+const SECONDS_IN_DAY = 86400;
+const SECONDS_IN_HOUR = 3600;
+const SECONDS_IN_MINUTE = 60;
+
+module.exports = {
+	ALREADY_CONNECTED_STATUS,
+	CANT_REACH_SNOWFLAKE_ERROR_STATUS,
+	CONNECTION_TIMED_OUT_CODE,
+	DEFAULT_CLIENT_APP_ID,
+	DEFAULT_CLIENT_APP_VERSION,
+	DEFAULT_WAREHOUSE,
+	DEFAULT_ROLE,
+	HACKOLADE_APPLICATION,
+	CLOUD_PLATFORM_POSTFIXES,
+	SECONDS_IN_DAY,
+	SECONDS_IN_HOUR,
+	SECONDS_IN_MINUTE,
+};
